requirements.txt
streamlit
transformers
torch
sentencepiece
pillow
pytesseract